add separate test
